+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - [#1564], [#1676], [#1697], [#1893] Display of DefineVideoStream tags with VLC player
 - List of treenode subitems on otherwise empty panel (with 32x32 icons)
 - DefineVideoStream codecId and videoFlagsDeblocking handled as enums in raw editation
+- Option to mute frame sounds
 
 ### Fixed
 - [#1897] Close menu button without selecting specific item